Remove duplicate S3 Gateway and also AWS from the name of the class
