+++ src/message-receiver/src/StoreInS3UseCase.integration.test.ts
@@ -1,12 +1,12 @@
 jest.retryTimes(10)
 import type { EventMessage } from "shared-types"
 import { isError } from "shared-types"
-import { TestAwsS3Gateway } from "shared"
+import { TestS3Gateway } from "shared"
 import type { StoreInS3Result } from "./StoreInS3UseCase"
 import StoreInS3UseCase from "./StoreInS3UseCase"
 import { auditLogEventsS3Config } from "shared-testing"
 
-const gateway = new TestAwsS3Gateway(auditLogEventsS3Config)
+const gateway = new TestS3Gateway(auditLogEventsS3Config)
 
 describe("StoreInS3UseCase", () => {
   beforeAll(async () => {
