--- src/shared/src/S3/index.ts
@@ -1,3 +1,0 @@
-export { default as AwsS3Gateway } from "./AwsS3Gateway"
-export { default as TestAwsS3Gateway } from "./TestAwsS3Gateway"
-export { default as createS3Config } from "./createS3Config"
