+++ src/shared/src/S3Gateway/index.ts
@@ -1,2 +1,3 @@
+export { default as createS3Config } from "./createS3Config"
 export { default as S3Gateway } from "./S3Gateway"
 export { default as TestS3Gateway } from "./TestS3Gateway"
