+++ src/shared/src/S3Gateway/parseGetObjectResponse.ts
