+++ src/shared/src/index.ts
@@ -2,6 +2,5 @@ export * from "./AuditLogApiClient"
 export * from "./BichardPostgresGateway"
 export * from "./Mq"
 export * from "./PostgresGateway"
-export * from "./S3"
 export * from "./S3Gateway"
 export * from "./utils"
